Fixed memory access - Using a larger type enable us to cover all possible memory

+++ core/memory.cpp
@@ -1,11 +1,11 @@
 #include "memory.h"
 
-uint8_t Memory::Read(uint8_t offset) const
+uint8_t Memory::Read(uint32_t offset) const
 {
     return m_data[offset];
 }
 
-void Memory::Write(uint8_t offset, uint8_t value)
+void Memory::Write(uint32_t offset, uint8_t value)
 {
     m_data[offset] = value;
 }
